update opendpi
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@29530 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 5b586a903a9cfe3ea96a2a0429bc7cac8c1bf74d

+++ src/router/opendpi/opendpi/lib/protocols/irc.c
@@ -507,7 +507,7 @@ static void ndpi_search_irc_tcp(struct ndpi_detection_module_struct *ndpi_struct
 		if (memcmp(packet->payload, "POST ", 5) == 0) {
 			ndpi_parse_packet_line_info(ndpi_struct, flow);
 			if (packet->parsed_lines) {
-				u_int16_t http_header_len = (packet_line(packet->parsed_lines - 1) - packet->payload) + 2;
+				u_int16_t http_header_len = (u_int16_t)((packet_line(packet->parsed_lines - 1) - packet->payload) + 2);
 				if (packet->payload_packet_len > http_header_len) {
 					http_content_ptr_len = packet->payload_packet_len - http_header_len;
 				}
